refactor: move ExtensionPhaseTiming value type into the audit contract
code_audit's report output held the last code_audit->extension type edge via
ExtensionPhaseTiming (a pure serde value type defined in
extension::runner_contract). Move it into homeboy-audit-contract alongside the
other audit value types and re-export it from extension::runner_contract so the
10 extension consumers stay unchanged, while code_audit imports it directly from
the contract.

Also repoints code_audit's remaining contract-owned fingerprint type imports
(HookRef, CallSite, DeadCodeMarker, UnusedParam, AggregateConstructionSeam,
AggregateLiteral) from the crate::extension re-export to homeboy_audit_contract
directly, matching the grammar/HookRef fixes in #8618/#8636.

Refs #8425
