+++ crates/homeboy-core/src/extension/runner_contract.rs
@@ -1,4 +1,4 @@
-use std::collections::{BTreeMap, HashSet};
+use std::collections::HashSet;
 
 use serde::{Deserialize, Serialize};
 
@@ -65,23 +65,12 @@ pub struct PhaseFailure {
     pub summary: String,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct ExtensionPhaseTiming {
-    pub name: String,
-    pub duration_ms: u64,
-    /// Provider-declared generic state for this phase, for example `running`,
-    /// `waiting`, `blocked`, `queued`, `passed`, or `failed`.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub status: Option<String>,
-    /// Human-readable provider summary for the phase. Core treats this as
-    /// opaque text and does not infer tool-specific behavior from it.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub message: Option<String>,
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub artifacts: Vec<serde_json::Value>,
-    #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-    pub metadata: BTreeMap<String, serde_json::Value>,
-}
+// `ExtensionPhaseTiming` is a pure serde value type that both the extension
+// runners (which produce it) and `code_audit` (which reports it) consume. It
+// lives in `homeboy-audit-contract` so `code_audit` never reaches up into the
+// extension layer for it; this re-export keeps the existing
+// `extension::runner_contract::ExtensionPhaseTiming` path working.
+pub use homeboy_audit_contract::ExtensionPhaseTiming;
 
 pub fn phase_status_from_exit_code(exit_code: i32) -> PhaseStatus {
     if exit_code == 0 {
